test: make createZustandContext a reusable util
